Moved email into a new file
